refactor: fixed types in hooks [2025-06-01]
