+++ gridwalk-ui/src/app/map/[workspaceId]/[mapName]/components/hooks/fileUploadUtils.ts
